Move useContactTraceData hook into its own file
